Don't depend on context.revisions anymore
We don't want to run the fetchRevisions hook in every command
that might need it (i.e. `deploy` and `activate`). Plugin authors
should implement a private function that fetches revisions if a
plugin needs that to function correctly.

See https://github.com/ember-cli/ember-cli-deploy/pull/209 for
a discussion about this.