Add support for previewing revisions
When using `hash`-location previewing revisions was already
possible. Users that are using `auto`-location can now opt-in
to using a provided instance-initializer that will rewrite
the ember apps baseURL while booting up so that the ember app
will boot up correctly and not fail with an unrecognized-url
error when providing a path like my-app.com/index.html:12345.