Add support for recusively listing all revisions in a bucket
`s3.listObjects` has a `MaxKeys` parameter which defaults to `1000`.
Unfortunately this also appears to be the highest value that you can
set it to. When you have more than 1000 items matching the prefix in
your bucket then some items were not being returned.

Since there is no sorting options this could result in the plugin
declaring "REVISION NOT FOUND!" when passed a value revision id.

This commit makes sure that we page through the results from the call
to `s3.listObjects` so that all results are loaded.