Add an option to disable URL encoding source
A bug in CEPH which is used as a drop in S3 replacement by some providers such as DigitalOcean, don't allow URL encoding of the `x-amz-copy-source` header.

https://tracker.ceph.com/issues/22121

The issue has been fixed but DigitalOcean still seems to have this issue and returns an "InvalidArgument" error when the `x-amz-copy-source` is URL encoded.

This adds an option (`urlEncodeSourceObject`) to disable URL encoding of the `x-amz-copy-source` header.