fix: Send `Prefix` to aws to limit files that we need to scan
Prior to this change, when you tried to activate a revision we list *all* revisions to
try and discover if the passed argument is valid. In our case this ends up fetching
tens of thousands of revisions from s3 which is kinda slow...

It turns out that the AWS `listObjects` function takes a prefix (this is already used
in the code to somewhat limit the results). We can pass the entire expected filename
and retrieve exactly one record (which is of course much much faster)