fix: Apply same optimisation to `upload` hook
We were again grabbing _every_ uploaded revision in order to be
able to check if you were trying to overwrite a revision. This is slow
when you have many revisions uploaded to s3.

Here we prefer our new `findRevision` method which uses the
`Prefix` to load information about _just_ the revision we're interested
in.