feat: Introduce an optional escape hatch for `listAllObjects`
There are cases where we might want to page through the list of objects available
on S3 but we don't need to keep going once we've found what we are looking for.
This commit updates `listAllObjects` to accept an `until` argument - if this
is provided then when a new page of results is loaded we check if `some` of them
match the `until` - if so we don't bother loading another page.

A concrete use-case for this is in #120 - sometimes we just want to
find the currently active revision so we only need to loop through the "pages"
on S3 until we do.