feat: Introduce a `getActiveRevision` method
This loads the head object and then calls `listAllObjects` with the new
`until` parameter and stops as soon as it has found an object which
matches the `ETag` of the active revision.

It returns the active revision in the same format that `fetchRevisions`
would - for backward compatibility with use-cases which were depending
on that.