feat: Add `fetchOnlyRelevantRevisions` config option when activating
If you set this option in your S3 config then we will only load enough revisions from
S3 to find the currently active version and then we will stop.
The assumption here is that the main use-case for `fetchInitialRevisions` is to be
able to do some sort of changelog or audit log from another plugin where we would
want to know the details of the previously active revision.
Since looping over revisions from S3 can be slow when you have too many of them
stored (see #120) we provide an option to short circuit that.