Handle empty s3 response
After the update to AWS SDK v3 when searching for a revision the
Contents on the response is not present if the revision hasn't been
created yet. This seemed like the smallest change to fix the problem,
and optional chaining is supported in Node 14, but I had to bump up the
ecmaVersion in the eslint config to allow it.